+++ seismiqb/src/__init__.py
@@ -1,16 +1,12 @@
 """ Init file. """
 # pylint: disable=wildcard-import
 # Core primitives
-from .cubeset import SeismicCubeset
+from .cubeset import SeismicCubeset, FaciesCubeset
 from .crop_batch import SeismicCropBatch
 
 # Data entities
 from .geometry import SeismicGeometry, BloscFile
-from .horizon import StructuredHorizon, Horizon
-from .horizon_unstructured import UnstructuredHorizon
-from .geobody import GeoBody
-from .fault import Fault
-from .facies import FaciesInfo, FaciesCubeset, Facies
+from .labels import Horizon, UnstructuredHorizon, Facies, Fault, GeoBody
 from .metrics import HorizonMetrics, GeometryMetrics
 from .samplers import GeometrySampler, HorizonSampler, SeismicSampler, BaseGrid, RegularGrid, ExtensionGrid
 
@@ -18,7 +14,5 @@ from .samplers import GeometrySampler, HorizonSampler, SeismicSampler, BaseGrid,
 # Utilities and helpers
 from .plotters import plot_image, plot_loss
 from .functional import *
-from .layers import *
 from .utils import *
-from .utility_classes import *
 from .controllers import *
