--- seismiqb/src/facies/__init__.py
@@ -1,4 +1,0 @@
-""" Facies-specific tasks classes. """
-from .info import FaciesInfo
-from .cubeset import FaciesCubeset
-from .facies import Facies
