+++ seismiqb/src/labels/__init__.py
@@ -1,0 +1,7 @@
+""" Labeled structures in a seismic volume. """
+from .horizon import Horizon
+from .horizon_facies import Facies
+from .horizon_unstructured import UnstructuredHorizon
+
+from .fault import Fault
+from .geobody import GeoBody
