+++ seismiqb/src/labels/fault_triangulation.py
