+++ seismiqb/src/labels/horizon_attributes.py
@@ -1,44 +1,23 @@
-""" Seismic facies container. """
+""" Mixin with computed along horizon geological attributes. """
 # pylint: disable=too-many-statements
-import os
-
 import numpy as np
-import pandas as pd
 from matplotlib import pyplot as plt
 
 from scipy.signal import hilbert, ricker
 from scipy.ndimage import convolve
+from scipy.ndimage.morphology import binary_fill_holes, binary_erosion
+from skimage.measure import label
 from sklearn.decomposition import PCA
 
-
+from ..functional import smooth_out
+from ..utils import to_list, retrieve_function_arguments, lru_cache
 from ..plotters import plot_image
-from ..horizon import Horizon
-from ..utils import to_list, retrieve_function_arguments
-from ..utility_classes import lru_cache
 
 
 
-class Facies(Horizon):
-    """ Extends base class functionality, allowing interaction with label subsets.
-
-    Parameters
-    ----------
-    subsets : dict, optional
-        Subsets of the initialized label. Keys are subset names and values are `Facies` instances.
-    args, kwargs : misc
-        For `Horizon` base class.
-
-    Methods
-    -------
-    - Class methods here rely heavily on the concept of nested subset storage. Facies are labeled along the horizon and
-    therefore can be viewed as subsets of those horizons, if compared as sets of triplets of points they consist of.
-
-    Main methods for interaction with label subsets are `add_subset` and `get_subset`. First allows adding given label
-    instance under provided name into parent subsets storage. Second returns the subset label under the requested name.
-
-
-    - If facies are added as subsets to their base horizons then their attributes can be accessed via the base label.
-    This is how `Facies` is different from `Horizon`, where different labels are stored in separate class attributes.
+class AttributesMixin:
+    """ Geological attributes along horizon.
+    !!.
 
     Method for getting desired attributes is `load_attribute`. It works with nested keys, i.e. one can get attributes
     of horizon subsets. Address method documentation for further details.
@@ -46,19 +25,7 @@ class Facies(Horizon):
 
     - Method `show` visualizes horizon and its attributes in both separate and overlap manners. It allows visual overlap
     of various attributes with one or more labels masks.
-
-
-    - Facies predictions might be stored as separate subsets of base labels (a horizon they are segmented along).
-    Method `evaluate` is used for comparison of ground truth labels and predicted ones, if both are stored as subsets.
-
-
-    - Since attributes calculation is cached, there may be a need to free up storage. Method `reset_caches` does that.
-
-
-    - To save facies call `dump`.
     """
-
-    # Correspondence between attribute alias and the class function that calculates it
     METHOD_TO_ATTRIBUTE = {
         'get_cube_values': ['cube_values', 'amplitudes'],
         'get_full_matrix': ['full_matrix', 'heights', 'depths'],
@@ -71,58 +38,334 @@ class Facies(Horizon):
     }
     ATTRIBUTE_TO_METHOD = {attr: func for func, attrs in METHOD_TO_ATTRIBUTE.items() for attr in attrs}
 
+    # Geometrical and geological properties
+    @property
+    def cube_values(self):
+        """ Values from the cube along the horizon. """
+        cube_values = self.get_cube_values(window=1)
+        cube_values[self.full_matrix == self.FILL_VALUE] = np.nan
+        return cube_values
 
-    def __init__(self, storage, geometry, name=None, dtype=np.int32, subsets=None, **kwargs):
-        super().__init__(storage=storage, geometry=geometry, name=name, dtype=dtype, **kwargs)
-        self.subsets = subsets or {}
+    @property
+    def amplitudes(self):
+        """ Alias for cube values. Depending on what loaded to cube geometries
+        might actually not be amplitudes, so use it with caution.
+        """
+        return self.cube_values
+
+    @property
+    def binary_matrix(self):
+        """ Matrix with ones at places where horizon is present and zeros everywhere else. """
+        return (self.matrix > 0).astype(bool)
+
+    @property
+    def borders_matrix(self):
+        """ Borders of horizons (borders of holes inside are not included). """
+        filled_matrix = self.filled_matrix
+        structure = np.ones((3, 3))
+        eroded = binary_erosion(filled_matrix, structure, border_value=0)
+        return filled_matrix ^ eroded # binary difference operation
+
+    @property
+    def boundaries_matrix(self):
+        """ Borders of horizons (borders of holes inside included). """
+        binary_matrix = self.binary_matrix
+        structure = np.ones((3, 3))
+        eroded = binary_erosion(binary_matrix, structure, border_value=0)
+        return binary_matrix ^ eroded # binary difference operation
+
+    @property
+    def coverage(self):
+        """ Ratio between number of present values and number of good traces in cube. """
+        return len(self) / (np.prod(self.cube_shape[:2]) - np.sum(self.geometry.zero_traces))
+
+    @property
+    def filled_matrix(self):
+        """ Binary matrix with filled holes. """
+        structure = np.ones((3, 3))
+        filled_matrix = binary_fill_holes(self.binary_matrix, structure)
+        return filled_matrix
+
+    @property
+    def full_matrix(self):
+        """ Matrix in cubic coordinate system. """
+        return self.get_full_matrix()
+
+    @property
+    def presence_matrix(self):
+        """ Binary matrix in cubic coordinate system. """
+        return self.put_on_full(self.binary_matrix, fill_value=False, dtype=bool)
+
+    @property
+    def grad_i(self):
+        """ Change of heights along iline direction. """
+        return self.grad_along_axis(0)
+
+    @property
+    def grad_x(self):
+        """ Change of heights along xline direction. """
+        return self.grad_along_axis(1)
+
+    @property
+    def hash(self):
+        """ Hash on current data of the horizon. """
+        return hash(self.matrix.data.tobytes())
+
+    @property
+    def horizon_metrics(self):
+        """ Calculate :class:`~HorizonMetrics` on demand. """
+        # pylint: disable=import-outside-toplevel
+        from ..metrics import HorizonMetrics
+        return HorizonMetrics(self)
+
+    @property
+    def instantaneous_phase(self):
+        """ Phase along the horizon. """
+        return self.horizon_metrics.evaluate('instantaneous_phase')
+
+    @property
+    def number_of_holes(self):
+        """ Number of holes inside horizon borders. """
+        holes_array = self.filled_matrix != self.binary_matrix
+        _, num = label(holes_array, connectivity=2, return_num=True, background=0)
+        return num
+
+    @property
+    def perimeter(self):
+        """ Number of points in the borders. """
+        return np.sum((self.borders_matrix == 1).astype(np.int32))
+
+    @property
+    def solidity(self):
+        """ Ratio of area covered by horizon to total area inside borders. """
+        return len(self) / np.sum(self.filled_matrix)
 
 
-    # Subsets interaction
-    def add_subset(self, name, item):
-        """ Add item to subsets storage.
+    # Carcass properties: should be used only if the horizon is a carcass
+    @property
+    def is_carcass(self):
+        """ Check if the horizon is a sparse carcass. """
+        return len(self) / self.filled_matrix.sum() < 0.5
+
+    @property
+    def carcass_ilines(self):
+        """ Labeled inlines in a carcass. """
+        uniques, counts = np.unique(self.points[:, 0], return_counts=True)
+        return uniques[counts > 256]
+
+    @property
+    def carcass_xlines(self):
+        """ Labeled xlines in a carcass. """
+        uniques, counts = np.unique(self.points[:, 1], return_counts=True)
+        return uniques[counts > 256]
+
+    @property
+    def carcass_grid(self):
+        """ Full matrix with present lines. """
+        return self.put_on_full(self.binary_matrix, fill_value=0.0)
+
+
+    # Retrieve data from seismic along horizon
+    def transform_where_present(self, array, normalize=None, fill_value=None, shift=None, rescale=None, res_ndim=None):
+        """ Normalize array where horizon is present, fill with constant where the horizon is absent.
 
         Parameters
         ----------
-        name : str
-            Key to store given horizon under.
-        item : Facies
-            Instance to store.
+        array : np.array
+            Matrix of (cube_ilines, cube_xlines, ...) shape.
+        normalize : 'min-max', 'mean-std', 'shift-rescale' or None/False
+            Normalization mode for data where `presence_matrix` is True.
+            If None, no normalization applied. Defaults to None.
+        fill_value : number
+            Value to fill `array` in where :attr:`.presence_matrix` is False. Must be compatible with `array.dtype`.
+            If None, no filling applied. Defaults to None.
+        shift, rescale : number, optional
+            For 'shift-rescale` normalization mode.
+        res_ndim : int or None
+            Number of dimensions returned result should have.
         """
-        self.subsets[name] = item
+        if not normalize:
+            pass
+        elif normalize == 'min-max':
+            values = array[self.presence_matrix]
+            min_, max_ = values.min(), values.max()
+            array = (array - min_) / (max_ - min_)
+        elif normalize == 'mean-std':
+            values = array[self.presence_matrix]
+            mean, std = values.mean(), values.std()
+            array = (array - mean) / std
+        elif normalize == 'shift-rescale':
+            array = (array + shift) * rescale
+        else:
+            raise ValueError('Unknown normalize mode `{}`'.format(normalize))
 
-    def get_subset(self, name):
-        """ Get item from subsets storage.
+        if fill_value is not None:
+            array[~self.presence_matrix] = fill_value
+
+        if res_ndim == array.ndim + 1:
+            array = array[..., np.newaxis]
+        elif res_ndim is not None and res_ndim != array.ndim:
+            msg = f"Result ndim is {array.ndim}, while requested ndim is {res_ndim}. "\
+                  f"Adding more than one new axis is not currently implemented."
+            raise ValueError(msg)
+
+        return array
+
+
+    @lru_cache(maxsize=1, apply_by_default=False, copy_on_return=True)
+    def get_cube_values(self, window=23, offset=0, chunk_size=256, **kwargs):
+        """ Get values from the cube along the horizon.
 
         Parameters
         ----------
-        name : str
-            Key desired item is stored under.
+        window : int
+            Width of data slice along the horizon.
+        offset : int
+            Offset of data slice with respect to horizon heights matrix.
+        chunk_size : int
+            Size of data along height axis processed at a time.
+        kwargs :
+            Passed directly to :meth:`.transform_where_present`.
         """
-        try:
-            return self.subsets[name]
-        except KeyError as e:
-            msg = f"Requested subset {name} is missing in subsets storage. Availiable subsets are {list(self.subsets)}."
-            raise KeyError(msg) from e
+        transform_kwargs = retrieve_function_arguments(self.transform_where_present, kwargs)
+        low = window // 2
+        high = max(window - low, 0)
+        chunk_size = min(chunk_size, self.h_max - self.h_min + window)
+
+        background = np.zeros((self.geometry.ilines_len, self.geometry.xlines_len, window), dtype=np.float32)
+
+        for h_start in range(max(low, self.h_min), self.h_max + 1, chunk_size):
+            h_end = min(h_start + chunk_size, self.h_max + 1)
+
+            # Get chunk from the cube (depth-wise)
+            location = (slice(None), slice(None),
+                        slice(h_start - low, min(h_end + high, self.geometry.depth)))
+            data_chunk = self.geometry.load_crop(location, use_cache=False)
+
+            # Check which points of the horizon are in the current chunk (and present)
+            idx_i, idx_x = np.asarray((self.matrix != self.FILL_VALUE) &
+                                      (self.matrix >= h_start) &
+                                      (self.matrix < h_end)).nonzero()
+            heights = self.matrix[idx_i, idx_x]
+
+            # Convert spatial coordinates to cubic, convert height to current chunk local system
+            idx_i += self.i_min
+            idx_x += self.x_min
+            heights -= (h_start - offset)
+
+            # Subsequently add values from the cube to background, then shift horizon 1 unit lower
+            for j in range(window):
+                background[idx_i, idx_x, np.full_like(heights, j)] = data_chunk[idx_i, idx_x, heights]
+                heights += 1
+                mask = heights < data_chunk.shape[2]
+                idx_i = idx_i[mask]
+                idx_x = idx_x[mask]
+                heights = heights[mask]
+
+        background[self.geometry.zero_traces == 1] = np.nan
+        return self.transform_where_present(background, **transform_kwargs)
+
+
+    @lru_cache(maxsize=1, apply_by_default=False, copy_on_return=True)
+    def get_full_matrix(self, **kwargs):
+        """ Transform `matrix` attribute to match cubic coordinates.
+
+        Parameters
+        ----------
+        kwargs :
+            Passed directly to :meth:`.put_on_full` and :meth:`.transform_where_present`.
+        """
+        transform_kwargs = retrieve_function_arguments(self.transform_where_present, kwargs)
+        matrix = self.put_on_full(self.matrix, **kwargs)
+        return self.transform_where_present(matrix, **transform_kwargs)
+
+
+    def get_array_values(self, array, shifts=None, grid_info=None, width=5, axes=(2, 1, 0)):
+        """ Get values from an external array along the horizon.
+
+        Parameters
+        ----------
+        array : np.ndarray
+            A data-array to make a cut from.
+        shifts : tuple or None
+            an offset defining the location of given array with respect to the horizon.
+            If None, `grid_info` with key `range` must be supplied.
+        grid_info : dict
+            Whenever passed, must contain key `range`.
+            Used for infering shifts of the array with respect to horizon.
+        width : int
+            required width of the resulting cut.
+        axes : tuple
+            if not None, axes-transposition with the required axes-order is used.
+        """
+        if shifts is None and grid_info is None:
+            raise ValueError('Either shifts or dataset with filled grid_info must be supplied!')
+
+        if shifts is None:
+            shifts = [grid_info['range'][i][0] for i in range(3)]
+
+        shifts = np.array(shifts)
+        horizon_shift = np.array((self.bbox[0, 0], self.bbox[1, 0]))
+
+        if axes is not None:
+            array = np.transpose(array, axes=axes)
+
+        # compute start and end-points of the ilines-xlines overlap between
+        # array and matrix in horizon and array-coordinates
+        horizon_shift, shifts = np.array(horizon_shift), np.array(shifts)
+        horizon_max = horizon_shift[:2] + np.array(self.matrix.shape)
+        array_max = np.array(array.shape[:2]) + shifts[:2]
+        overlap_shape = np.minimum(horizon_max[:2], array_max[:2]) - np.maximum(horizon_shift[:2], shifts[:2])
+        overlap_start = np.maximum(0, horizon_shift[:2] - shifts[:2])
+        heights_start = np.maximum(shifts[:2] - horizon_shift[:2], 0)
+
+        # recompute horizon-matrix in array-coordinates
+        slc_array = [slice(l, h) for l, h in zip(overlap_start, overlap_start + overlap_shape)]
+        slc_horizon = [slice(l, h) for l, h in zip(heights_start, heights_start + overlap_shape)]
+        overlap_matrix = np.full(array.shape[:2], fill_value=self.FILL_VALUE, dtype=np.float32)
+        overlap_matrix[slc_array] = self.matrix[slc_horizon]
+        overlap_matrix -= shifts[-1]
+
+        # make the cut-array and fill it with array-data located on needed heights
+        result = np.full(array.shape[:2] + (width, ), np.nan, dtype=np.float32)
+        iterator = [overlap_matrix + shift for shift in range(-width // 2 + 1, width // 2 + 1)]
+
+        for i, surface_level in enumerate(np.array(iterator)):
+            mask = (surface_level >= 0) & (surface_level < array.shape[-1]) & (surface_level !=
+                                                                               self.FILL_VALUE - shifts[-1])
+            mask_where = np.where(mask)
+            result[mask_where[0], mask_where[1], i] = array[mask_where[0], mask_where[1],
+                                                            surface_level[mask_where].astype(np.int)]
+
+        return result
+
+
+    # Modify things
+    def grad_along_axis(self, axis=0):
+        """ Change of heights along specified direction. """
+        grad = np.diff(self.matrix, axis=axis, prepend=0)
+        grad[np.abs(grad) > 10000] = self.FILL_VALUE
+        grad[self.matrix == self.FILL_VALUE] = self.FILL_VALUE
+        return grad
+
+    def make_float_matrix(self, kernel=None, kernel_size=7, sigma=2., margin=5, iters=1):
+        """ Smooth the depth matrix to produce floating point numbers. """
+        float_matrix = smooth_out(self.full_matrix, kernel=kernel,
+                                  kernel_size=kernel_size, sigma=sigma, margin=margin,
+                                  fill_value=self.FILL_VALUE, preserve=True, iters=iters)
+        return float_matrix
+
+    def put_on_full(self, matrix=None, fill_value=None, dtype=np.float32):
+        """ Create a matrix in cubic coordinate system. """
+        matrix = matrix if matrix is not None else self.matrix
+        fill_value = fill_value if fill_value is not None else self.FILL_VALUE
+
+        background = np.full(self.cube_shape[:-1], fill_value, dtype=dtype)
+        background[self.i_min:self.i_max+1, self.x_min:self.x_max+1] = matrix
+        return background
 
 
     # Generic attributes loading
-    def _load_attribute(self, src, location=None, use_cache=True, **kwargs):
-        """ Load horizon attribute at requested location. """
-        try:
-            method_name = self.ATTRIBUTE_TO_METHOD[src]
-        except KeyError as e:
-            raise ValueError(f"Unknown `src` {src}. Expected one of {list(self.ATTRIBUTE_TO_METHOD.keys())}.") from e
-
-        method = getattr(self, method_name)
-        data = method(use_cache=use_cache, **kwargs)
-        # TODO: Someday, we would need to re-write attribute loading methods
-        # so they use locations not to crop the loaded result, but to load attribute only at location.
-        if location is not None:
-            i_slice, x_slice, _ = location
-            data = data[i_slice, x_slice]
-
-        return data
-
     def load_attribute(self, src, location=None, **kwargs):
         """ Load horizon or its subset attribute values at requested location.
 
@@ -176,6 +419,23 @@ class Facies(Horizon):
             # pylint: disable=protected-access
             mask = subset._load_attribute(src='masks', location=location, fill_value=0).astype(bool)
             data[~mask] = kwargs.get('fill_value', self.FILL_VALUE)
+
+        return data
+
+    def _load_attribute(self, src, location=None, use_cache=True, **kwargs):
+        """ Load horizon attribute at requested location. """
+        try:
+            method_name = self.ATTRIBUTE_TO_METHOD[src]
+        except KeyError as e:
+            raise ValueError(f"Unknown `src` {src}. Expected one of {list(self.ATTRIBUTE_TO_METHOD.keys())}.") from e
+
+        method = getattr(self, method_name)
+        data = method(use_cache=use_cache, **kwargs)
+        # TODO: Someday, we would need to re-write attribute loading methods
+        # so they use locations not to crop the loaded result, but to load attribute only at location.
+        if location is not None:
+            i_slice, x_slice, _ = location
+            data = data[i_slice, x_slice]
 
         return data
 
@@ -333,7 +593,7 @@ class Facies(Horizon):
 
 
     # Attributes visualization
-    def show(self, attributes='depths', mode='imshow', return_figure=False, **kwargs):
+    def show(self, attributes='depths', mode='imshow', return_figure=False, enlarge=True, width=9, **kwargs):
         """ Display facies attributes with predefined defaults.
 
         Loads requested data, constructs default parameters wrt to that data and delegates plot to `plot_image`.
@@ -403,7 +663,13 @@ class Facies(Horizon):
             data = self.load_attribute(**load)
             return postprocess(data)
 
+        def enlarge_data(data):
+            if self.is_carcass and enlarge:
+                data = self.enlarge_carcass_image(data, width)
+            return data
+
         data = apply_by_scenario(load_data, attributes)
+        data = apply_by_scenario(enlarge_data, data)
 
         # Make titles
         def extract_data_name(attributes):
@@ -478,76 +744,3 @@ class Facies(Horizon):
         plt.show()
 
         return figure if return_figure else None
-
-
-    # Metrics evaluations
-    def evaluate(self, src_true, src_pred, metrics_fn, metrics_names=None, output='df'):
-        """ Apply given function to 'masks' attribute of requested labels.
-
-        Parameters
-        ----------
-        src_true : str
-            Name of `labels` subset to load true mask from.
-        src_pred : str
-            Name of `labels` subset to load prediction mask from.
-        metrics_fn : callable or list of callable
-            Metrics function(s) to calculate.
-        metrics_name : str, optional
-            Name of the column with metrics values in resulted dataframe.
-        output : 'df' or 'arr'
-            Whether return an array of metrics values or dataframe.
-        """
-        pd.options.display.float_format = '{:,.3f}'.format
-
-        labeled_traces = self.get_full_binary_matrix(fill_value=0).astype(bool)
-        true = self.load_attribute(f"{src_true}/masks", fill_value=0)[labeled_traces]
-        pred = self.load_attribute(f"{src_pred}/masks", fill_value=0)[labeled_traces]
-
-        metrics_fn = to_list(metrics_fn)
-        values = [fn(true, pred) for fn in metrics_fn]
-
-        if output == 'arr':
-            return values
-
-        index = pd.MultiIndex.from_arrays([[self.geometry.displayed_name], [self.short_name]],
-                                          names=['geometry_name', 'horizon_name'])
-        names = metrics_names if metrics_names is not None else [fn.__name__ for fn in metrics_fn]
-        df = pd.DataFrame(index=index, data=values, columns=names)
-        return df
-
-
-    # Matrix inversion
-    def __sub__(self, other):
-        if not isinstance(other, type(self)):
-            raise TypeError(f"Operands types do not match. Got {type(self)} and {type(other)}.")
-        presence = other.presence_matrix
-        discrepancies = self.full_matrix[presence] != other.full_matrix[presence]
-        if discrepancies.any():
-            raise ValueError("Horizons have different depths where present.")
-        result = self.full_matrix.copy()
-        result[presence] = self.FILL_VALUE
-        name = f"~{other.name}"
-        return type(self)(result, self.geometry, name)
-
-    def invert_subset(self, subset):
-        """ Subtract subset matrix from facies matrix. """
-        return self - self.get_subset(subset)
-
-
-    # Manage data
-
-    def dump(self, path, name=None, log=True):
-        """ Save facies. """
-        path = path.replace('*', self.geometry.short_name)
-        name = name.replace('*', self.name) if name is not None else self.name
-        os.makedirs(path, exist_ok=True)
-        dump_path = f"{path}/{name}"
-        super().dump(dump_path)
-        if log:
-            print(f"`{self.short_name}` saved to `{dump_path}`")
-
-    def reset_cache(self):
-        """ Clear cached data. """
-        super().reset_cache()
-        for subset_label in self.subsets.values():
-            subset_label.reset_cache()
