+++ seismiqb/src/utils/__init__.py
@@ -1,0 +1,5 @@
+""" Utility functions and classes. """
+# pylint: disable=wildcard-import
+from .classes import *
+from .functions import *
+from .layers import *
